feat: add treeleaf strategy
Adds `treeleaf` strategy to put all non-leaf nodes in the first CAR, and then all leaf nodes in subsequent CARs.

Note: no attempt is made to restrict the tree CAR size... if you have a huge tree and a smol target size, it may well exceeed it... this PR is a sketch of the idea of the "put all the tree meta data in one CAR, and the actual file data (leaf nodes) in other cars.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>